feat: add preprocessing (#4)
* feat: feature selection init

* fix: boxplots displaying incorrectly (#2)

re-added matplotlib

* feat: preprocessing init

added fill_na function

* feat: added bind data function

* feat: added remove unneeded floats func

* feat: added get dummies func

* chore: added fill na docstring

* feat: fill_na function now supports train, test

left original function commented as new version is untested

* fix: scale_data now accepts train and test dfs

* feat: add brand spelling fix func

* chore: add docstring to fix_brands functions

* feat: fix models functions

this function does not work 100%, might need to replace/add a knn on top

* feat: fix transmission functions

* feat: fix fuel type functions

* fix: fix_model_spelling works with null brands

* feat: fix no brand models function

fills null brand with brand name when model is not null

* fix: no brand models issue with ambiguous models

ambiguous models are hardcoded as such, this function does not work like the other spell fixing ones, might change in the future for consistency

* chore: notes about models

* feat: feature selection init

* feat: preprocessing init

added fill_na function

* feat: added bind data function

* feat: added remove unneeded floats func

* feat: added get dummies func

* chore: added fill na docstring

* feat: fill_na function now supports train, test

left original function commented as new version is untested

* fix: scale_data now accepts train and test dfs

* feat: add brand spelling fix func

* chore: add docstring to fix_brands functions

* feat: fix models functions

this function does not work 100%, might need to replace/add a knn on top

* feat: fix transmission functions

* feat: fix fuel type functions

* fix: fix_model_spelling works with null brands

* feat: fix no brand models function

fills null brand with brand name when model is not null

* fix: no brand models issue with ambiguous models

ambiguous models are hardcoded as such, this function does not work like the other spell fixing ones, might change in the future for consistency

* chore: notes about models

* refactor: preprocessing

all fix functions with the exception of fix_models now use the polars expr api instead of map_elements. this commit includes changes to comply with ruff rules.

* chore: update vscode workspace settings

* refactor: fill_na

* style: fix type hinting

* fix: fix get_dummies not supporting train test

* chore: remove specific fix_data functions

prefer using the generic function with the old params

* feat: add drop_columns function